+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## Unreleased-->
+## Unreleased
+
+### Added
+
+- add `message_id` to `Email` and webhook `EmailBody` types (https://github.com/resend/resend-rust/pull/77)
+
 
 ## [0.28.0] - 2026-07-05