+++ CHANGELOG.md
@@ -6,13 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+<!--## Unreleased-->
+
+## [0.28.0] - 2026-07-27
 
 ### Added
 
 - add `message_id` to `Email` and webhook `EmailBody` types (https://github.com/resend/resend-rust/pull/77)
 - `suppressions` endpoints
-
 
 ## [0.28.0] - 2026-07-05
 
@@ -562,6 +563,7 @@ Disabled `reqwest`'s default features and enabled `rustls-tls`.
 
 Initial release.
 
+[0.29.0]: https://crates.io/crates/resend-rs/0.29.0
 [0.28.0]: https://crates.io/crates/resend-rs/0.28.0
 [0.27.0]: https://crates.io/crates/resend-rs/0.27.0
 [0.26.1]: https://crates.io/crates/resend-rs/0.26.1